增加根据输入选择topk的逻辑（已简单验证精度）

+++ vllm_ascend/distributed/P2PAFDConnector.py
@@ -109,11 +109,9 @@ class P2PAFDConnector(AFDConnectorBase):
 
             ae_group.send(topk_weights_size)            
             ae_group.send(topk_weights)
-            # print(topk_weights_size, topk_weights.size())
 
             ae_group.send(topk_ids_size)
             ae_group.send(topk_ids)
-            # print(topk_ids_size, topk_ids.size())
         return
 
     # MOE发给ATTN（ATTN接收）hidden_states只负责提供shape和dtype
@@ -146,13 +144,11 @@ class P2PAFDConnector(AFDConnectorBase):
 
             topk_weights_size = ae_group.recv(2,dtype=torch.int64)
             topk_weights = torch.empty([topk_weights_size[0],topk_weights_size[1]])
-            topk_weights = ae_group.recv(topk_weights.size(),dtype=torch.float32)
-            topk_weights = topk_weights.to(torch.bfloat16)
+            topk_weights = ae_group.recv(topk_weights.size(),dtype=torch.bfloat16)
 
             topk_ids_size = ae_group.recv(2,dtype=torch.int64)
             topk_ids = torch.empty([topk_ids_size[0],topk_ids_size[1]])
             topk_ids = ae_group.recv(topk_ids.size(),dtype=torch.int32)
-            # print(topk_ids_size, topk_ids.size())
 
             ffn_need_metadata_obj.topk_weights = topk_weights
             ffn_need_metadata_obj.topk_ids = topk_ids
